requirements.txt
# Core AI Agent Framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AI Model Providers
anthropic>=0.34.0
openai>=1.0.0
boto3>=1.34.0  # For AWS Bedrock if needed

# Model Context Protocol (MCP) Servers
mcp>=1.0.0

# File Operations and Memory Management
pyyaml>=6.0.1
python-dateutil>=2.8.2
pathlib2>=2.3.7

# Environment and Configuration Management
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI Interface
click>=8.1.7
rich>=13.7.0  # For enhanced terminal output
prompt-toolkit>=3.0.43  # For interactive CLI

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development Tools
black>=23.12.0  # Code formatting
flake8>=6.1.0   # Linting
mypy>=1.8.0     # Type checking

# Utilities
requests>=2.31.0
aiofiles>=23.2.1  # Async file operations
uvloop>=0.19.0    # Fast event loop (Unix only)

# Memory and Search
chromadb>=0.4.0   # For vector embeddings if needed
sentence-transformers>=2.2.0  # For semantic search

# Logging and Monitoring
structlog>=23.2.0 